+++ src/app/page.tsx
@@ -1,9 +1,5 @@
 import Home from '@/app/_component/page/home';
 
 export default function Index() {
-  return (
-    <main>
-      <Home />
-    </main>
-  );
+  return <Home />;
 }
