+++ src/app/post/[slug]/page.tsx
@@ -1,38 +1,10 @@
 import { Metadata } from 'next';
 import { notFound } from 'next/navigation';
-import { getAllPosts, getPostBySlug } from '@/lib/api';
-import { CMS_NAME } from '@/lib/constants';
-import markdownToHtml from '@/lib/markdownToHtml';
-import Container from '@/app/_component/template/container';
-import Header from '@/app/_component/organism/header';
-import { PostBody } from '@/app/_component/organism/post-body';
-import { PostHeader } from '@/app/_component/organism/post-header';
+import PostPage from '@/app/_component/page/post';
+import { getPostBySlug } from '@/lib/api';
 
 export default async function Post({ params }: Params) {
-  const post = getPostBySlug(params.slug);
-
-  if (!post) {
-    return notFound();
-  }
-
-  const content = await markdownToHtml(post.content || '');
-
-  return (
-    <main>
-      <Container>
-        <Header />
-        <article className='mb-32'>
-          <PostHeader
-            title={post.title}
-            coverImage={post.coverImage}
-            date={post.date}
-            author={post.author}
-          />
-          <PostBody content={content} />
-        </article>
-      </Container>
-    </main>
-  );
+  return <PostPage slug={params.slug} />;
 }
 
 type Params = {
@@ -48,7 +20,7 @@ export function generateMetadata({ params }: Params): Metadata {
     return notFound();
   }
 
-  const title = `${post.title} | Next.js Blog Example with ${CMS_NAME}`;
+  const title = `${post.title} | Next.js Blog Example with markdown`;
 
   return {
     title,
@@ -57,12 +29,4 @@ export function generateMetadata({ params }: Params): Metadata {
       images: [post.ogImage.url],
     },
   };
-}
-
-export async function generateStaticParams() {
-  const posts = getAllPosts();
-
-  return posts.map((post) => ({
-    slug: post.slug,
-  }));
 }
